Refactor node deletion
